util improvements and new results

+++ util.py
@@ -211,14 +211,17 @@ if __name__ == "__main__":
 					db.plot(selected_keys)
 				action_selection_menu(db, selected_keys)
 			elif i == len(keys)+2:
-				s = ''
-				for key in selected_keys:
-					s+=str(key)+' '
-				print 'Removing keys {}'.format(s)
-				answer = raw_input('Confirm (yes)? ')
-				if answer == 'yes':
-					db.remove(selected_keys)
-					action_selection_menu(db, [])
+				if len(selected_keys) > 0:
+					s = ''
+					for key in selected_keys:
+						s+=str(key)+' '
+					print 'Removing keys {}'.format(s)
+					answer = raw_input('Confirm (yes)? ')
+					if answer == 'yes':
+						db.remove(selected_keys)
+						action_selection_menu(db, [])
+					else:
+						action_selection_menu(db, selected_keys)
 				else:
 					action_selection_menu(db, selected_keys)
 			elif i == len(keys)+3:
